chore: split satosa backend methods in file per scopes
